Split channels tests to one per class to handle parameters variable sharing and guarantee clean up.
